Remove duplicate AuthContext and update wallet login
